@@ -14,48 +14,51 @@
     "@istanbuljs/nyc-config-typescript": "1.0.1",
     "@types/mocha": "8.0.3",
     "@types/mockdate": "2.0.0",
-    "@types/node": "14.11.2",
-    "@types/sinon": "9.0.5",
-    "@types/sinon-express-mock": "1.3.8",
+    "@types/node": "14.14.6",
+    "@types/sinon": "9.0.8",
+    "@types/sinon-express-mock": "1.3.9",
     "@types/supertest": "2.0.10",
-    "@typescript-eslint/eslint-plugin": "4.2.0",
-    "@typescript-eslint/parser": "4.2.0",
+    "@typescript-eslint/eslint-plugin": "4.6.0",
+    "@typescript-eslint/parser": "4.6.0",
     "check-node-version": "4.0.3",
     "dotenv": "8.2.0",
-    "eslint": "7.9.0",
+    "eslint": "7.12.1",
     "eslint-config-airbnb-base": "14.2.0",
-    "eslint-config-prettier": "6.11.0",
+    "eslint-config-prettier": "6.15.0",
     "eslint-config-recommended": "4.1.0",
-    "eslint-plugin-import": "2.22.0",
+    "eslint-plugin-import": "2.22.1",
     "eslint-plugin-prettier": "3.1.4",
     "eslint-plugin-promise": "4.2.1",
+    "git-revision-webpack-plugin": "3.0.6",
     "husky": "4.3.0",
     "improved-yarn-audit": "2.3.1",
-    "jest": "26.4.2",
+    "jest": "26.6.1",
     "lerna": "3.22.1",
     "mocha": "8.1.3",
     "mochawesome": "6.1.1",
     "mockdate": "3.0.2",
     "node-rsa": "1.1.1",
-    "nodemon": "2.0.4",
+    "nodemon": "2.0.6",
     "nyc": "15.1.0",
     "prettier": "2.1.2",
     "should": "13.2.3",
-    "sinon": "9.0.3",
+    "sinon": "9.2.1",
     "sinon-express-mock": "2.2.1",
     "source-map-support": "0.5.19",
-    "supertest": "4.0.2",
-    "ts-jest": "26.4.0",
-    "ts-loader": "8.0.4",
-    "ts-mocha": "7.0.0",
+    "supertest": "6.0.0",
+    "ts-jest": "26.4.3",
+    "ts-loader": "8.0.7",
+    "ts-mocha": "8.0.0",
     "ts-node": "9.0.0",
     "tsconfig-paths": "3.9.0",
     "typedoc": "0.19.2",
     "typedoc-plugin-lerna-packages": "0.3.1",
-    "typescript": "4.0.3",
+    "typescript": "4.0.5",
     "unit.js": "2.1.1",
-    "webpack": "4.44.2",
-    "webpack-cli": "3.3.12"
+    "webpack": "5.3.2",
+    "webpack-cli": "4.1.0",
+    "webpack-merge": "5.3.0",
+    "wtfnode": "0.8.3"
   },
   "scripts": {
     "build": "yarn run:concurrent build",
