@@ -13,14 +13,14 @@
     "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'",
     "test:unit": "exit 0"
   },
-  "dependencies": {
+  "devDependencies": {
     "@types/git-revision-webpack-plugin": "3.0.0",
-    "@types/webpack": "4.41.22",
-    "@types/webpack-merge": "4.1.5",
-    "git-revision-webpack-plugin": "3.0.6",
-    "webpack": "4.44.2",
-    "webpack-merge": "5.1.4",
-    "wrapper-webpack-plugin": "2.1.0"
+    "@types/webpack-merge": "4.1.5"
+  },
+  "peerDependencies": {
+    "git-revision-webpack-plugin": "3.0.x",
+    "webpack": "5.3.x",
+    "webpack-merge": "5.2.x"
   },
   "keywords": [
     "mds"
