@@ -1,9 +1,12 @@
 /*
     Copyright 2019 City of Los Angeles.
+    
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
+
         http://www.apache.org/licenses/LICENSE-2.0
+    
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -11,7 +14,17 @@
     limitations under the License.
  */
 
+import logger from '@mds-core/mds-logger'
+import cache from '@mds-core/mds-agency-cache'
 import { ApiServer, HttpServer } from '@mds-core/mds-api-server'
 import { api } from './api'
 
-HttpServer(ApiServer(api), { port: process.env.COMPLIANCE_API_HTTP_PORT })
+cache
+  .startup()
+  .then(() => {
+    return HttpServer(ApiServer(api), { port: process.env.COMPLIANCE_API_HTTP_PORT })
+  })
+  // eslint-disable-next-line promise/prefer-await-to-callbacks
+  .catch(err => {
+    logger.error('mds-compliance startup failure', err)
+  })
